update environment (#118)
* update environment

* update python version

* downgrade numpy

* downgrade primesieve

* downgrade numpy

* downgrade everything

* adding python versions

* adding python versions

* remove python versions

* attempt to update numpy version

* removing python 3.8

* update primesieve version

* remove the use of primesieve package

* fix some solutions

* fix more solutions

* remove the sieve function as implemented in python

* add more python versions

* remove py 3.9, 3.10 and 3.11

+++ solutions/PE132.py
@@ -12,7 +12,7 @@ Solve time: ~0.05 seconds
 """
 from util.utils import timeit
 import unittest
-from primesieve import primes
+from util.utils import primes_upto as primes
 
 
 # R(n) = (10^n - 1) / 9
@@ -31,7 +31,7 @@ class Problem132:
         count = 0
         ls_primes = primes(max_prime)
         for p in ls_primes:
-            if pow(10, n, 9 * p) == 1:
+            if pow(10, n, int(9 * p)) == 1:
                 count += 1
                 ans += p
             if count == max_count:
